added localized strings, comments for configuration class, code refactoring
