+++ TextFieldsCatalog/Resources/Strings/en.lproj/Localizable.strings
@@ -23,6 +23,8 @@
 "TextFieldType.Bordered.description" = "• Границы поля ввода скруглены и подсвечены.\n• Имеется плейсхолдер над полем ввода.\n• Информационное-сообщение или сообщение об ошибке внизу, в одну строку.\n• Кастомизируется под ввод пароля.\n• Поддержка валидаторов и форматтеров.";
 "TextFieldType.Underlined.title" = "Подчеркнутое поле ввода";
 "TextFieldType.Underlined.description" = "• Под полем ввода присутствует линия.\n• Плейсхолдер выполнен как CATextLayer, что позволяет анимационно изменять его кегль, цвет и позицию при переходе между состояниями.\n• Кастомизируется под ввод пароля.\n• Информационное-сообщение или сообщение об ошибке внизу, в одну строку.\n• Поддержка валидаторов и форматтеров.";
+"TextFieldType.CustomUnderlined.title" = "Кастомизированое поле ввода с подчеркивание";
+"TextFieldType.CustomUnderlined.description" = "• То же самое, что и поле ввода с подчеркиванием, но кастомизированное.\n• Отсутствует лейбл с подсказкой/ошибкой.";
 
 // MARK: - Presets
