+++ TextFieldsCatalog/TextFields/TextFields/Configuration/TextFieldConfigurations.swift
@@ -75,123 +75,3 @@ final class BorderedTextFieldConfiguration {
 //        background = BackgroundConfiguration()
 //    }
 }
-
-final class LineConfiguration {
-    let smallHeight: CGFloat
-    let bigHeight: CGFloat
-    let cornerRadius: CGFloat
-    let colors: ColorConfiguration
-
-    init(smallHeight: CGFloat,
-         bigHeight: CGFloat,
-         cornerRadius: CGFloat,
-         colors: ColorConfiguration) {
-        self.smallHeight = smallHeight
-        self.bigHeight = bigHeight
-        self.cornerRadius = cornerRadius
-        self.colors = colors
-    }
-}
-
-final class FloatingPlaceholderConfiguration {
-    let font: UIFont
-    let topPosition: CGRect
-    let bottomPosition: CGRect
-    let smallFontSize: CGFloat
-    let bigFontSize: CGFloat
-    let topColors: ColorConfiguration
-    let bottomColors: ColorConfiguration
-
-    init(font: UIFont,
-         topPosition: CGRect,
-         bottomPosition: CGRect,
-         smallFontSize: CGFloat,
-         bigFontSize: CGFloat,
-         topColors: ColorConfiguration,
-         bottomColors: ColorConfiguration) {
-        self.font = font
-        self.topPosition = topPosition
-        self.bottomPosition = bottomPosition
-        self.smallFontSize = smallFontSize
-        self.bigFontSize = bigFontSize
-        self.topColors = topColors
-        self.bottomColors = bottomColors
-    }
-}
-
-final class PlaceholderConfiguration {
-
-}
-
-final class TextFieldConfiguration {
-    let font: UIFont
-    let defaultPadding: UIEdgeInsets
-    let increasedPadding: UIEdgeInsets
-    let tintColor: UIColor
-    let colors: ColorConfiguration
-
-    init(font: UIFont,
-         defaultPadding: UIEdgeInsets,
-         increasedPadding: UIEdgeInsets,
-         tintColor: UIColor,
-         colors: ColorConfiguration) {
-        self.font = font
-        self.defaultPadding = defaultPadding
-        self.increasedPadding = increasedPadding
-        self.tintColor = tintColor
-        self.colors = colors
-    }
-}
-
-final class TextFieldBorderConfiguration {
-
-}
-
-final class HintConfiguration {
-    let font: UIFont
-    let colors: ColorConfiguration
-
-    init(font: UIFont, colors: ColorConfiguration) {
-        self.font = font
-        self.colors = colors
-    }
-}
-
-final class PasswordModeConfiguration {
-    let secureModeOnImage: UIImage
-    let secureModeOffImage: UIImage
-    let normalColor: UIColor
-    let pressedColor: UIColor
-
-    init(secureModeOnImage: UIImage,
-         secureModeOffImage: UIImage,
-         normalColor: UIColor,
-         pressedColor: UIColor) {
-        self.secureModeOnImage = secureModeOnImage
-        self.secureModeOffImage = secureModeOffImage
-        self.normalColor = normalColor
-        self.pressedColor = pressedColor
-    }
-}
-
-final class BackgroundConfiguration {
-    let color: UIColor
-
-    init(color: UIColor) {
-        self.color = color
-    }
-}
-
-final class ColorConfiguration {
-    let error: UIColor
-    let normal: UIColor
-    let active: UIColor
-    let disabled: UIColor
-
-    init(error: UIColor, normal: UIColor, active: UIColor, disabled: UIColor) {
-        self.error = error
-        self.normal = normal
-        self.active = active
-        self.disabled = disabled
-    }
-}
